FISH-10327 Some cleanup
Deletes some unused properties, cleans up the arquillian extensions, and adds starting & stopping of the Derby database
