+++ ejb-platform-tck/src/test/java/fish/payara/arquillian/PayaraTckExtension.java
@@ -1,0 +1,12 @@
+package fish.payara.arquillian;
+
+import org.jboss.arquillian.core.spi.LoadableExtension;
+import org.jboss.arquillian.test.spi.enricher.resource.ResourceProvider;
+
+public class PayaraTckExtension implements LoadableExtension {
+    @Override
+    public void register(ExtensionBuilder builder) {
+        builder.service(ResourceProvider.class, PayaraXmlProcessor.class);
+        builder.observer(PayaraXmlProcessor.class);
+    }
+}
