--- ejb-platform-tck/src/test/java/wildfly/arquillian/JBossTckExtension.java
@@ -1,13 +1,0 @@
-package wildfly.arquillian;
-
-import org.jboss.arquillian.container.test.spi.client.deployment.ApplicationArchiveProcessor;
-import org.jboss.arquillian.core.spi.LoadableExtension;
-import org.jboss.arquillian.test.spi.enricher.resource.ResourceProvider;
-
-public class JBossTckExtension implements LoadableExtension {
-    @Override
-    public void register(ExtensionBuilder builder) {
-        builder.service(ResourceProvider.class, JBossXmlProcessor.class);
-        builder.observer(JBossXmlProcessor.class);
-    }
-}
