+++ ejb-platform-tck/src/test/resources/META-INF/services/org.jboss.arquillian.core.spi.LoadableExtension
@@ -1,1 +1,1 @@
-wildfly.arquillian.JBossTckExtension
+fish.payara.arquillian.PayaraTckExtension
