--- ejb-platform-tck/src/test/resources/jakartaeetck/tmp/tstest.jte
@@ -1,88 +1,0 @@
-#Properties for test: autoTimerWithoutInfo
-#Tue Apr 08 14:23:13 BST 2025
-Driver=DriverOnlyUsedIfDriverManagerIsYes
-DriverManager=no
-authpassword=javajoe
-authuser=javajoe
-binarySize=24
-cofSize=5
-cofTypeSize=5
-db.dml.file=derby/derby.dml.sql
-db.supports.sequence=true
-db1=db1OnlyUsedIfDriverManagerIsYes
-db2=db1OnlyUsedIfDriverManagerIsYes
-finder=cts
-ftable=ctstable2
-harness.log.delayseconds=0
-harness.log.port=2000
-harness.log.traceflag=false
-harness.socket.retry.count=10
-harness.temp.directory=${ts.home}/tmp
-imap.port=143
-iofile=MailIOFile.txt
-java.naming.factory.initial=foo
-javaee.level=full
-javamail.mailbox=test1
-javamail.password=cts1
-javamail.protocol=imap
-javamail.root.path=""
-javamail.server=jsepc04.us.oracle.com
-javamail.username=cts1@jsepc04.us.oracle.com
-jdbc.db=derby
-jimage.dir=${ts.home}/tmp/jdk-bundles
-jms_timeout=10000
-jstl.db.password=${derby.passwd}
-jstl.db.user=${derby.user}
-log.file.location=${vi.domain}/logs
-logical.hostname.servlet=server
-longvarbinarySize=50
-mailuser1=cts1@jsepc04.us.oracle.com
-optional.tech.packages.to.ignore=jakarta.xml.bind
-org.omg.CORBA.ORBClass=foo
-password=j2ee
-password1=cts1
-platform.mode=jakartaEE
-porting.ts.HttpsURLConnection.class.1=com.sun.ts.lib.implementation.sun.javaee.SunRIHttpsURLConnection
-porting.ts.HttpsURLConnection.class.2=com.sun.ts.lib.implementation.sun.javaee.SunRIHttpsURLConnection
-porting.ts.deploy.class.1=com.sun.ts.lib.implementation.sun.javaee.glassfish.AutoDeployment
-porting.ts.deploy.class.2=com.sun.ts.lib.implementation.sun.javaee.glassfish.AutoDeploymentSeparateVM
-porting.ts.jms.class.1=com.sun.ts.lib.implementation.sun.javaee.SunRIJMSAdmin
-porting.ts.jms.class.2=com.sun.ts.lib.implementation.sun.javaee.SunRIJMSAdmin
-porting.ts.login.class.1=com.sun.ts.lib.implementation.sun.javaee.GlassFishLoginContext
-porting.ts.login.class.2=com.sun.ts.lib.implementation.sun.javaee.GlassFishLoginContext
-porting.ts.url.class.1=com.sun.ts.lib.implementation.sun.common.SunRIURL
-porting.ts.url.class.2=com.sun.ts.lib.implementation.sun.common.SunRIURL
-ptable=ctstable1
-rapassword1=cts1
-rapassword2=cts2
-rauser1=cts1
-rauser2=cts2
-s1as.modules=${javaee.home}/modules
-securedWebServicePort=1044
-service_eetest.vehicles=ejbliteservlet
-sigTestClasspath=${s1as.modules}/glassfish-corba-omgapi.jar${pathsep}${s1as.modules}/glassfish-corba-orb.jar${pathsep}${s1as.modules}/jakarta.enterprise.cdi-api.jar${pathsep}${s1as.modules}/jakarta.json-api.jar${pathsep}${s1as.modules}/jakarta.json.bind-api.jar${pathsep}${s1as.modules}/jakarta.batch-api.jar${pathsep}${s1as.modules}/jakarta.interceptor-api.jar${pathsep}${s1as.modules}/stax2-api.jar${pathsep}${s1as.modules}/jakarta.enterprise.concurrent-api.jar${pathsep}${s1as.modules}/jakarta.websocket-api.jar${pathsep}${s1as.modules}/jakarta.websocket-client-api.jar${pathsep}${s1as.modules}/jakarta.jms-api.jar${pathsep}${s1as.modules}/jakarta.faces.jar${pathsep}${s1as.modules}/jakarta.validation-api.jar${pathsep}${s1as.modules}/jakarta.annotation-api.jar${pathsep}${s1as.modules}/jakarta.xml.bind-api.jar${pathsep}${s1as.modules}/webservices-api-osgi.jar${pathsep}${pathsep}${s1as.modules}/jakarta.ws.rs-api.jar${pathsep}${s1as.modules}/weld-osgi-bundle.jar${pathsep}${s1as.modules}/jakarta.ejb-api.jar${pathsep}${s1as.modules}/jakarta.mail-api.jar${pathsep}${s1as.modules}/jakarta.persistence-api.jar${pathsep}${s1as.modules}/jakarta.resource-api.jar${pathsep}${s1as.modules}/jakarta.authorization-api.jar${pathsep}${s1as.modules}/jakarta.authentication-api.jar${pathsep}${s1as.modules}/jakarta.servlet-api.jar${pathsep}${s1as.modules}/jakarta.inject-api.jar${pathsep}${s1as.modules}/jakarta.el-api.jar${pathsep}${s1as.modules}/jakarta.servlet.jsp-api.jar${pathsep}${s1as.modules}/jakarta.servlet.jsp.jstl-api.jar${pathsep}${jtaJarClasspath}${pathsep}${s1as.modules}/jakarta.security.enterprise-api.jar${pathsep}${s1as.modules}/jakarta.activation-api.jar${pathsep}${jimage.dir}/java.base${pathsep}${jimage.dir}/java.rmi${pathsep}${jimage.dir}/java.sql${pathsep}${jimage.dir}/java.naming
-smtp.port=25
-test_classname=com.sun.ts.tests.ejb32.lite.timer.schedule.auto.attr.singleton.Client
-transport_protocol=smtp
-ts_home=dummy
-user=j2ee
-user1=cts1
-varbinarySize=48
-variable.mapper=org.glassfish.expressly.lang.VariableMapperImpl
-vehicle=ejbliteservlet
-vehicle_archive_name=ejb32_lite_timer_schedule_auto_attr_singleton_ejbliteservlet_vehicle_web
-webServerHost=localhost
-webServerPort=8080
-whitebox-anno_no_md=java\:comp/env/eis/whitebox-anno_no_md
-whitebox-mdcomplete=java\:comp/env/eis/whitebox-mdcomplete
-whitebox-mixedmode=java\:comp/env/eis/whitebox-mixedmode
-whitebox-multianno=java\:comp/env/eis/whitebox-multianno
-whitebox-notx=java\:comp/env/eis/whitebox-notx
-whitebox-notx-param=java\:comp/env/eis/whitebox-notx-param
-whitebox-permissiondd=java\:comp/env/eis/whitebox-permissiondd
-whitebox-tx=java\:comp/env/eis/whitebox-tx
-whitebox-tx-param=java\:comp/env/eis/whitebox-tx-param
-whitebox-xa=java\:comp/env/eis/whitebox-xa
-whitebox-xa-param=java\:comp/env/eis/whitebox-xa-param
-work.dir=/files/JTWork
-ws_wait=5
